# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "covhub 控制 API",
    "description": "JaCoCo 运行期覆盖率的采集与看板。整套方案只有 hub 这一个服务端，被测机器和发版节点不装 Python、不装 java、不放配置文件，全部通过这些接口驱动 hub 干活 —— 它们只需要 curl。\n\n**写接口在 hub 内部串行执行**，返回体里带着这次执行的日志。**非 2xx 一律表示失败**，调用方应据此让部署流程停下来。\n\n发版节点上更省事的做法是用 integration/covhub-client.sh 包一层。",
    "version": "2.0.0"
  },
  "paths": {
...
    "/api/status": {
      "get": {
        "tags": [
          "查询"
        ],
        "summary": "连通性与最新覆盖率",
        "description": "看板上那些数字的 JSON 版。push 服务的 unknown=true 表示当前进程没有收集端，\n说不出在线与否 —— 那不等于离线。",
        "operationId": "status_api_status_get",
        "security": [
          {
            "tokenHeader": []
          },
          {
            "tokenQuery": []
          },
          {
            "tokenCookie": []
          }
        ],
        "parameters": [
          {
            "name": "service",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "只看某一个服务；不给则全部",
              "title": "Service"
            },
            "description": "只看某一个服务；不给则全部"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusResponse"
                }
              }
            }
          },
          "401": {
            "description": "令牌无效或缺失",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "没有这个服务",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "hub 内部错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/agent-opts": {
      "get": {
        "tags": [
          "查询"
        ],
        "summary": "取该服务应注入的 -javaagent 参数串",
        "description": "被测服务零侵入接入的入口：把返回的参数串塞进 JAVA_TOOL_OPTIONS 即可，\n不改代码不改 pom。加 format=text 直接出纯文本，方便 shell 里 $(curl ...)。",
        "operationId": "agent_opts_api_agent_opts_get",
        "security": [
          {
            "tokenHeader": []
          },
          {
            "tokenQuery": []
          },
          {
            "tokenCookie": []
          }
        ],
        "parameters": [
          {
            "name": "service",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "服务名，须与 hub 配置一致",
              "title": "Service"
            },
            "description": "服务名，须与 hub 配置一致"
          },
          {
            "name": "format",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "填 text 则返回纯文本而不是 JSON",
              "title": "Format"
            },
            "description": "填 text 则返回纯文本而不是 JSON"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentOpts"
                }
              },
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "令牌无效或缺失",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "没有这个服务",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "hub 内部错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/agent.jar": {
      "get": {
        "tags": [
          "查询"
        ],
        "summary": "下载 jacocoagent.jar",
        "description": "被测机器不必预先铺一份 agent，容器的 initContainer 一条 curl 就能拿到。\n读的是配置里 jacocoAgent 指向的文件 —— 那一项填的是**被测端**路径，\n两者不一致时这个接口会 404，但 agent-opts 照样输出正确的参数串。",
        "operationId": "agent_jar_api_agent_jar_get",
        "responses": {
          "200": {
            "description": "jar 文件",
            "content": {
              "application/json": {
                "schema": {}
              },
              "application/java-archive": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "401": {
            "description": "令牌无效或缺失",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "配置里 jacocoAgent 指向的文件在 hub 上不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "hub 内部错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "tokenHeader": []
          },
          {
            "tokenQuery": []
          },
          {
            "tokenCookie": []
          }
        ]
      }
    },
    "/api/diagnose": {
      "get": {
        "tags": [
          "查询"
        ],
        "summary": "诊断 exec 与 class 产物是否对得上",
        "description": "**任何覆盖率数字不对劲，先跑这个。** 它把 exec 里记录的 class id 和 classfiles\n的 class id 求交集 —— 匹配率低就是 class 产物对不上，这是接入时最贵、最难查、\n而且**不会报错**的一个坑（报告只会显示全部未覆盖）。",
        "operationId": "diagnose_api_diagnose_get",
        "security": [
          {
            "tokenHeader": []
          },
          {
            "tokenQuery": []
          },
          {
            "tokenCookie": []
          }
        ],
        "parameters": [
          {
            "name": "service",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "服务名",
              "title": "Service"
            },
            "description": "服务名"
          },
          {
            "name": "version",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "诊断某个已归档版本，不给则诊断当前周期",
              "title": "Version"
            },
            "description": "诊断某个已归档版本，不给则诊断当前周期"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiagnoseResponse"
                }
              }
            }
          },
          "401": {
            "description": "令牌无效或缺失",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "没有这个服务",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "hub 内部错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/dump": {
      "post": {
        "tags": [
          "采集"
        ],
        "summary": "拉一次快照并出报告",
        "description": "累加，不清零。采完会重新渲染看板。",
        "operationId": "dump_api_dump_post",
        "security": [
          {
            "tokenHeader": []
          },
          {
            "tokenQuery": []
          },
          {
            "tokenCookie": []
          }
        ],
        "parameters": [
          {
            "name": "service",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "服务名",
              "title": "Service"
            },
            "description": "服务名"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandResult"
                }
              }
            }
          },
          "401": {
            "description": "令牌无效或缺失",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "没有这个服务",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "hub 内部错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "业务上失败了（目标不可达、没有数据……），log 里有原因",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/report": {
      "post": {
        "tags": [
          "采集"
        ],
        "summary": "用已有 exec 重出报告",
        "description": "不碰被测服务，只是拿现有数据重新生成一次报告 —— 改了 reportExcludes 之后用。",
        "operationId": "report_api_report_post",
        "security": [
          {
            "tokenHeader": []
          },
          {
            "tokenQuery": []
          },
          {
            "tokenCookie": []
          }
        ],
        "parameters": [
          {
            "name": "service",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "服务名",
              "title": "Service"
            },
            "description": "服务名"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandResult"
                }
              }
            }
          },
          "401": {
            "description": "令牌无效或缺失",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "没有这个服务",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "hub 内部错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "业务上失败了（目标不可达、没有数据……），log 里有原因",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/predeploy": {
      "post": {
        "tags": [
          "发版"
        ],
        "summary": "结算并归档（停服前必须调用）",
        "description": "dump --reset + 归档 + 出终版报告。\n\n**必须在停服之前执行。** 服务一停 agent 随进程消失，那段覆盖率没有任何补救手段\n—— 所以目标不可达时它**故意报 409**，好让部署流程停下来。确实要跳过就带\nallowMissing=1，但那意味着这段数据已经丢了。",
        "operationId": "predeploy_api_predeploy_post",
        "security": [
          {
            "tokenHeader": []
          },
          {
            "tokenQuery": []
          },
          {
            "tokenCookie": []
          }
        ],
        "parameters": [
          {
            "name": "service",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "服务名",
              "title": "Service"
            },
            "description": "服务名"
          },
          {
            "name": "version",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "版本标识，不给则取配置里的 version",
              "title": "Version"
            },
            "description": "版本标识，不给则取配置里的 version"
          },
          {
            "name": "allowMissing",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "填 1 则目标已离线时不报错",
              "title": "Allowmissing"
            },
            "description": "填 1 则目标已离线时不报错"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandResult"
                }
              }
            }
          },
          "401": {
            "description": "令牌无效或缺失",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "没有这个服务",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "hub 内部错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "业务上失败了（目标不可达、没有数据……），log 里有原因",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AgentOpts": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok",
            "default": true
          },
          "service": {
            "type": "string",
            "title": "Service"
          },
          "agentOpts": {
            "type": "string",
            "title": "Agentopts",
            "description": "塞进被测服务 JAVA_TOOL_OPTIONS 的参数串",
            "examples": [
              "-javaagent:/opt/jacoco/jacocoagent.jar=output=tcpserver,address=0.0.0.0,port=6300,includes=com.example.*,sessionid=1.4.2"
            ]
          }
        },
        "type": "object",
        "required": [
          "service",
          "agentOpts"
        ],
        "title": "AgentOpts"
      },
...
      "DiagnoseResponse": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok",
            "default": true
          },
          "diagnose": {
            "$ref": "#/components/schemas/Diagnose"
          }
        },
        "type": "object",
        "required": [
          "diagnose"
        ],
        "title": "DiagnoseResponse"
      },
      "Error": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok",
            "default": false
          },
          "error": {
            "type": "string",
            "title": "Error",
            "description": "给人看的失败原因"
          },
          "hint": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hint"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "Error"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "StatusResponse": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok",
            "default": true
          },
          "services": {
            "items": {
              "$ref": "#/components/schemas/ServiceStatus"
            },
            "type": "array",
            "title": "Services"
          }
        },
        "type": "object",
        "required": [
          "services"
        ],
        "title": "StatusResponse"
      },
...
    },
    "securitySchemes": {
      "tokenHeader": {
        "type": "apiKey",
        "description": "推荐。令牌走请求头而不是 URL，免得被 access log 和 ps 输出记下来",
        "in": "header",
        "name": "X-Covhub-Token"
      },
      "tokenQuery": {
        "type": "apiKey",
        "description": "浏览器里访问看板时用；hub 会种一个 Cookie 再跳回不带令牌的地址",
        "in": "query",
        "name": "token"
      },
      "tokenCookie": {
        "type": "apiKey",
        "description": "?token= 带对一次之后由 hub 自动种下，浏览器后续请求靠它",
        "in": "cookie",
        "name": "covhub_token"
      }
    }
  },
...
}